Extract saga contract types into shared libs/atlas-saga library
Consolidates duplicated saga types (Saga, Step, Action, Status, payloads,
validation conditions) from 10 services into a new shared library. Each
service now re-exports shared types via aliases for backward compatibility.
Removes the monolithic libs/atlas-script-core/saga/model.go (910 lines).

+++ services/atlas-character-factory/atlas.com/character-factory/factory/processor.go
@@ -167,7 +167,7 @@ func buildCharacterCreationOnlySaga(transactionId uuid.UUID, input RestModel) sa
 		Mp:           input.Mp,
 		Face:         input.Face,
 		Hair:         input.Hair + input.HairColor,
-		Skin:         uint32(input.SkinColor),
+		Skin:         input.SkinColor,
 		Top:          input.Top,
 		Bottom:       input.Bottom,
 		Shoes:        input.Shoes,
